Update aprendizados.json
As I mentioned in my bio, I started my career in a telemarketing job and after almost a year, I got a call from my old boss. He came up to me and asked, “Do you like IT?” I thought about leaving the operation. If I said yes, then yes. I was tired of answering the phone and would be happy with anything they offered me because it was the improvement I had chosen for most of my life up until that point…